requirements.txt
absl-py @ file:///home/conda/feedstock_root/build_artifacts/absl-py_1751547525079/work
astunparse @ file:///home/conda/feedstock_root/build_artifacts/astunparse_1736248061654/work
biopython @ file:///home/conda/feedstock_root/build_artifacts/biopython_1761734597652/work
Brotli @ file:///home/conda/feedstock_root/build_artifacts/brotli-split_1761591974641/work
cached-property @ file:///home/conda/feedstock_root/build_artifacts/cached_property_1615209429212/work
certifi @ file:///home/conda/feedstock_root/build_artifacts/certifi_1759648874697/work/certifi
cffi @ file:///home/conda/feedstock_root/build_artifacts/cffi_1725560558132/work
charset-normalizer @ file:///home/conda/feedstock_root/build_artifacts/charset-normalizer_1760437218288/work
contourpy @ file:///home/conda/feedstock_root/build_artifacts/contourpy_1756544597990/work
cycler @ file:///home/conda/feedstock_root/build_artifacts/cycler_1733332471406/work
flatbuffers @ file:///home/conda/feedstock_root/build_artifacts/python-flatbuffers_1758880282135/work
fonttools @ file:///home/conda/feedstock_root/build_artifacts/fonttools_1759187078489/work
gast @ file:///home/conda/feedstock_root/build_artifacts/gast_1733299680490/work
google-pasta @ file:///home/conda/feedstock_root/build_artifacts/google-pasta_1733852499742/work
grpcio @ file:///home/conda/feedstock_root/build_artifacts/grpc-split_1761049705887/work
h2 @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_h2_1756364871/work
h5py @ file:///home/conda/feedstock_root/build_artifacts/h5py_1760616433023/work
hpack @ file:///home/conda/feedstock_root/build_artifacts/hpack_1737618293087/work
hyperframe @ file:///home/conda/feedstock_root/build_artifacts/hyperframe_1737618333194/work
idna @ file:///home/conda/feedstock_root/build_artifacts/idna_1760286409563/work
importlib_metadata @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_importlib-metadata_1747934053/work
joblib @ file:///home/conda/feedstock_root/build_artifacts/joblib_1756321760188/work
keras @ file:///home/conda/feedstock_root/build_artifacts/keras_1761616457014/work
kiwisolver @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_kiwisolver_1756467527/work
lxml==6.0.2
Markdown @ file:///home/conda/feedstock_root/build_artifacts/markdown_1757093412127/work
markdown-it-py @ file:///home/conda/feedstock_root/build_artifacts/markdown-it-py_1754951200865/work
MarkupSafe @ file:///home/conda/feedstock_root/build_artifacts/markupsafe_1759055168201/work
matplotlib==3.10.7
mdurl @ file:///home/conda/feedstock_root/build_artifacts/mdurl_1733255585584/work
ml-dtypes @ file:///home/conda/feedstock_root/build_artifacts/ml_dtypes_1756742168282/work
munkres==1.1.4
namex @ file:///home/conda/feedstock_root/build_artifacts/namex_1748346368839/work
networkx @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_networkx_1749078300/work
numpy @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_numpy_1761161698/work/dist/numpy-2.3.4-cp312-cp312-linux_x86_64.whl#sha256=79c666d07103be5d9b660f1e9e31acb1feac4b5915c9ed27da1fc67423c046b5
opt_einsum @ file:///home/conda/feedstock_root/build_artifacts/opt_einsum_1733687912731/work
optree @ file:///home/conda/feedstock_root/build_artifacts/optree_1756812157611/work
packaging @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_packaging_1745345660/work
pandas @ file:///home/conda/feedstock_root/build_artifacts/pandas_1759265541279/work
pandas-stubs==2.3.2.250926
pillow @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_pillow_1758208668/work
protobuf @ file:///home/conda/feedstock_root/build_artifacts/protobuf_1760392535057/work/bazel-bin/python/dist/protobuf-6.31.1-cp312-abi3-linux_x86_64.whl#sha256=5beb26a8f0e4500239c1c215ad814622cad863a91fa5b0ace3e4acac6ed12179
pycparser @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_pycparser_1733195786/work
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1750615794071/work
pyparsing @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_pyparsing_1758436411/work
PySide6==6.8.2
PySocks @ file:///home/conda/feedstock_root/build_artifacts/pysocks_1733217236728/work
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_python-dateutil_1751104122/work
pytz @ file:///home/conda/feedstock_root/build_artifacts/pytz_1742920838005/work
PyYAML @ file:///home/conda/feedstock_root/build_artifacts/pyyaml_1758891829620/work
requests @ file:///home/conda/feedstock_root/build_artifacts/requests_1755614211359/work
rich @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_rich_1760026188/work/dist
scikit-learn @ file:///home/conda/feedstock_root/build_artifacts/scikit-learn_1757406006820/work/dist/scikit_learn-1.7.2-cp312-cp312-linux_x86_64.whl#sha256=129fb53912f867c66de404c45e47a17c8d2466e12961a80dadee213c841e7450
scipy @ file:///home/conda/feedstock_root/build_artifacts/scipy-split_1761689805279/work/dist/scipy-1.16.3-cp312-cp312-linux_x86_64.whl#sha256=99192edb305485cb50d0365378a3f86e5776626d1565153a1a3c2d4be32f8b87
setuptools==80.9.0
shiboken6==6.8.2
six @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_six_1753199211/work
spektral==1.3.1
tensorboard @ file:///home/conda/feedstock_root/build_artifacts/tensorboard_1740522445235/work/tensorboard-2.19.0-py3-none-any.whl#sha256=5e71b98663a641a7ce8a6e70b0be8e1a4c0c45d48760b076383ac4755c35b9a0
tensorboard_data_server @ file:///home/conda/feedstock_root/build_artifacts/tensorboard-data-server_1759413095605/work/tensorboard_data_server-0.7.0-py3-none-manylinux2014_x86_64.whl#sha256=73e1d5805423850ce08c1c0e1d01cf0566a54b8b3fc3345c8eaede8d95e269bb
tensorflow @ file:///home/conda/feedstock_root/build_artifacts/tensorflow-split_1761488409304/work/tensorflow_pkg/tensorflow-2.19.0-cp312-cp312-linux_x86_64.whl#sha256=7330f895eb6bc48d46610ec7d893f3fab8c12cab9d075fcf5e7bc59eae7fbf5d
tensorflow_estimator @ file:///home/conda/feedstock_root/build_artifacts/tensorflow-split_1761488409304/work/tensorflow-estimator/wheel_dir3.12/tensorflow_estimator-2.15.0-py2.py3-none-any.whl#sha256=9f87f56fc4d9a2524dd07af21652608dc21d010e5cfe4c6805aa42b96fc47ef5
termcolor @ file:///home/conda/feedstock_root/build_artifacts/termcolor_1761565758678/work
threadpoolctl @ file:///home/conda/feedstock_root/build_artifacts/threadpoolctl_1741878222898/work
tornado @ file:///home/conda/feedstock_root/build_artifacts/tornado_1756854920986/work
tqdm==4.67.1
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_typing_extensions_1756220668/work
tzdata @ file:///home/conda/feedstock_root/build_artifacts/python-tzdata_1742745135198/work
unicodedata2 @ file:///home/conda/feedstock_root/build_artifacts/unicodedata2_1756494453070/work
urllib3 @ file:///home/conda/feedstock_root/build_artifacts/urllib3_1750271362675/work
Werkzeug @ file:///home/conda/feedstock_root/build_artifacts/werkzeug_1733160440960/work
wheel==0.45.1
wrapt @ file:///home/conda/feedstock_root/build_artifacts/wrapt_1760964140464/work
zipp @ file:///home/conda/feedstock_root/build_artifacts/zipp_1749421620841/work
zstandard==0.25.0
